requirements.txt
# Django related packages
Django==3.2.16 # Can update to pip install Django==4.1.3 
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
django-rest-auth==0.9.5
django-extensions==3.2.0
django-cors-headers==3.13.0
django-allauth==0.51.0

# Core API
coreapi==2.3.3

# Other
python-dotenv==0.20.0